requirements.txt
Flask
numpy
tensorflow
requests
gunicorn
Pillow
